Fix: Accept bytecode of length 65536 bytes
In order to support the filter bytecode maximum length (65536 bytes),
the lttng_ust_filter_bytecode len field type must be able to hold more
than a uint16_t. Change the field type to a uint32_t.

Also, since the relocation table is located at the end of the actual
bytecode, the reloc_table_offset (reloc_offset in ust-abi) field must
support offset values larger than 65535. Change the field type to a
uint32_t. This change will allow support of relocation table appended to
larger bytecode without breaking the ABI if the need arise in the
future.

Both changes currently breaks the filter ABI, but this should be a
reasonable compromise since the filtering feature has not been released
yet.

Acked-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>

+++ src/lib/lttng-ctl/filter/filter-visitor-generate-bytecode.c
@@ -239,7 +239,7 @@ int visit_node_load(struct filter_parser_ctx *ctx, struct ir_op *node)
 		uint32_t insn_len = sizeof(struct load_op)
 			+ sizeof(struct field_ref);
 		struct field_ref ref_offset;
-		uint16_t reloc_offset;
+		uint32_t reloc_offset;
 
 		insn = calloc(insn_len, 1);
 		if (!insn)
